+++ server/migrations/index.js
@@ -31,8 +31,9 @@ import { migrateToV60 } from './v60RuntimePluginStates.js'
 import { migrateToV61 } from './v61EvolutionEvidence.js'
 import { migrateToV62 } from './v62EvolutionExclusions.js'
 import { migrateToV63 } from './v63EvolutionCandidates.js'
+import { migrateToV64 } from './v64EvolutionReplay.js'
 
-export { migrateToV31, migrateToV32, migrateToV33, migrateToV34, migrateToV35, migrateToV36, migrateToV37, migrateToV38, migrateToV39, migrateToV40, migrateToV41, migrateToV42, migrateToV43, migrateToV44, migrateToV45, migrateToV46, migrateToV47, migrateToV48, migrateToV49, migrateToV50, migrateToV51, migrateToV52, migrateToV53, migrateToV54, migrateToV55, migrateToV56, migrateToV57, migrateToV58, migrateToV59, migrateToV60, migrateToV61, migrateToV62, migrateToV63 }
+export { migrateToV31, migrateToV32, migrateToV33, migrateToV34, migrateToV35, migrateToV36, migrateToV37, migrateToV38, migrateToV39, migrateToV40, migrateToV41, migrateToV42, migrateToV43, migrateToV44, migrateToV45, migrateToV46, migrateToV47, migrateToV48, migrateToV49, migrateToV50, migrateToV51, migrateToV52, migrateToV53, migrateToV54, migrateToV55, migrateToV56, migrateToV57, migrateToV58, migrateToV59, migrateToV60, migrateToV61, migrateToV62, migrateToV63, migrateToV64 }
 
 /**
  * V2-V30 stay in db.js for upgrade compatibility. New migrations are registered
@@ -72,6 +73,7 @@ export const schemaMigrations = Object.freeze([
   { version: 61, up: migrateToV61 },
   { version: 62, up: migrateToV62 },
   { version: 63, up: migrateToV63 },
+  { version: 64, up: migrateToV64 },
 ])
 
 export const LATEST_SCHEMA_VERSION = schemaMigrations.at(-1)?.version || 30